Check isActive state locally in components
